+++ microbit_v2/src/animations/multi_color_strand.rs
@@ -69,7 +69,7 @@ impl Animation for MultiColorStrand<'_> {
         }
 
         ws2812.write(self.data.borrow().iter().copied()).unwrap();
-        timer.delay_ms(settings.delay());
+        timer.delay_ms(settings.delay() as u32);
     }
 
     fn reset(&mut self) {
