+++ microbit_v2/src/animations/uni_color_fade_in.rs
@@ -52,7 +52,7 @@ impl Animation for UniColorFadeIn<'_> {
         }
 
         ws2812.write(self.data.borrow().iter().copied()).unwrap();
-        timer.delay_ms(settings.delay());
+        timer.delay_ms(settings.delay() as u32);
     }
 
     fn reset(&mut self) {
